+++ impulse-client-kit/components/src/lib.rs
@@ -23,6 +23,7 @@ pub mod cookie_consent;
 pub mod data_table;
 pub mod date_picker;
 pub mod dialog;
+pub mod dismiss;
 pub mod drawer;
 pub mod dropdown_menu;
 pub mod empty;